src/myapi/requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
PyJWT==2.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
mysqlclient==2.2.0
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SQLAlchemy==2.0.23
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.1
flask-cors==4.0.0
python-dotenv==1.0.0
pypdf==3.17.2
fpdf==1.7.2
